Add basic unit test coverage for Python client (MR !1926)
* Adds some mising Python bindings like UNSPEC enum values
* Fix RANGE not being masked to 20 bits when parsing packets in Python
* Fix different __str__ behavior for enums using newest pybind11:
  https://github.com/pybind/pybind11/issues/2537
* Fix digest module using platform-specific bit widths for header fields
  causing spurious test failures when checking hashes

Approved-by: Kairen Wong
Approved-by: Chris Bayruns
Approved-by: Pavlo Bashmakov

+++ python/src/ouster/client/_digest.py
@@ -81,10 +81,12 @@ class ScanDigest:
         hashes = {}
         hashes.update({c.name: _md5(ls.field(c)) for c in ChanField})
 
-        hashes['TIMESTAMP'] = _md5(np.array([h.timestamp for h in ls.headers]))
+        hashes['TIMESTAMP'] = _md5(
+            np.array([h.timestamp for h in ls.headers], dtype=np.uint64))
         hashes['ENCODER_COUNT'] = _md5(
-            np.array([h.encoder for h in ls.headers]))
-        hashes['VALID'] = _md5(np.array([h.status for h in ls.headers]))
+            np.array([h.encoder for h in ls.headers], dtype=np.uint64))
+        hashes['VALID'] = _md5(
+            np.array([h.status for h in ls.headers], dtype=np.uint64))
 
         return cls(**hashes)
